viperaveil-main/Discord/requirements.txt
#moviepy==1.0.3
#pygifsicle==1.0.7
aiohttp==3.8.3
aiosqlite==0.18.0
asgiref==3.6.0
asyncio==3.4.3
asyncpg==0.27.0
azure-cognitiveservices-vision-computervision==0.9.0
beautifulsoup4==4.11.2
bs4==0.0.1
Flask-Compress==1.13
Flask-Dance==6.2.0
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Migrate==4.0.0
flask-restx==1.0.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
Flask[async]==2.1.3
git+https://github.com/theGunner295/soundcloud-lib.git
nest-asyncio==1.5.6
pillow==9.5
psutil==5.9.4
psycopg2==2.9.5
py-cord==2.3.2
pydub==0.25.1
PyNaCl==1.5.0
python-decouple==3.7
requests==2.28.1
sponsorblock.py==0.2.2
spotify==0.10.2
sqlalchemy-orm==1.2.4
SQLAlchemy==1.4.44
tekore==4.5.0
waitress==2.1.2
wavelink==1.3.5
youtube-dl==2021.12.17
youtube-search-python==1.6.6